pyproject.toml
[tool.poetry]
name = "main_site"
version = "0.1.0"
description = ""
authors = ["Consultora Criativa - contato@consultoracriativa.design"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.19.2"
django-grappelli = "^2.15.4"
django-import-export = "^2.7.1"
django-debug-toolbar = "^3.2.4"
psycopg2-binary = "^2.9.3"
Pillow = "^9.0.0"
Django = "^4.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
